Customized TRANSFER transaction for BigchainDB.
A helper can now provide and update personal information (a self-introduction), which is stored in BigchainDB. This is done by performing a TRANSFER operation. The asset representing the helper is transferred back to himself/herself with metadata, which contains content of personal information, updated.

Users classes were modified. Variables were encapsulated in classes for easier management.

Also, the actions in AppDriver.java was made cancellable.