Added a section in the literature survey for blockchain databases.
The title of the section was changed to "Blockchain databases" (the original title: "Blockchain for data sharing"). Uncertain parts that are subject to changes were denoted with "<-- (?) -->". Fixed an image display issue.

Sadly, I may have overdone the literature survey. It may be trimmed down later, but I believe this literature survey is informative.